Import RHEL's test Sanity/testsuite
It's a wrapper for the upstream test suite.

There's one small addition (already inherited from RHEL) covering
but I'm keeping it here for convenience as it works with the upstream
test suite too. I've slightly modified it to make it work beyond
RHEL 8. The modification is based on the final upstream fix [1] and
the related discussion in thread [2].

[1] https://sourceware.org/git/gitweb.cgi?p=binutils-gdb.git;h=6bf4a34047452f882c5cc66bd85812ee1bb5a41c
[2] https://sourceware.org/pipermail/binutils/2018-August/104304.html