Update the Sanity/testsuite test for new build location
6bfd06790befd11d1c11be000a159aa51c4b2e1b changed where the build
actually happens, so these cp commands fail and break the test.
This change is based on the version in
https://gitlab.com/redhat/centos-stream/rpms/binutils/-/blob/c9s/tests/Sanity/testsuite/runtest.sh
but reimplemented in line with the rest of this somewhat
different version of the test (I am not sure what the story is
with these different versions of the tests in Fedora vs. CentOS
Stream, I'm trying to find that out).

Signed-off-by: Adam Williamson <awilliam@redhat.com>