Fix BuildRequires on bison
bison is only brought in for gold architectures or when
building gprofng right now, but non-gold architectures need
it too if the source file for the parser happens to have been
touched. Make the dependency unconditional to ensure that the
package can always be built.

While at it, clean up a few related things.

A BuildRequires on gcc-c++ already exists a few lines below
for the non-clang case, and obviously we don't need it when
building with clang.

m4 is not used directly by us, but rather by bison. Luckily
the bison package correctly depends on m4, so we don't need
to bring it in ourselves.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>